Expand the installation instructions (#23)
* Expand the installation instructions

Since Strada depends on Stimulus and imports "@hotwired/stimulus" in its
own source code, we need to make sure that the import is available.

The easiest way to do this is to use importmap-rails, which will
automatically add the importmap script tag to your layout. If you're not
using importmap-rails, you can manually add the importmap script tag to
your layout, or use a JS bundler such as Webpack or esbuild.

* Remove controller registration from install instructions

It depends on how you're using Stimulus. For example, with stimulus-rails
you don't need to register controllers manually.